% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractions.R
\docType{data}
\name{attractions}
\alias{attractions}
\title{Attraction meta data}
\format{
A data frame with 87 rows and 8 variables:
\describe{
\item{code}{Code for attraction, alphanumeric}
\item{name}{Disney's official attraction name, alphanumeric}
\item{short_name}{Short name, alphanumeric}
\item{state}{State park is located in (Flordia or California), alphanumeric}
\item{park}{Disney theme park where the attraction is located}
\item{land}{Section of the park an attraction is located}
\item{opened_on}{Attraction's opening date, date}
\item{duration}{Duration of attraction in minutes, numeric}
\item{average_wait_per_hundred}{Hypothetical wait time for every 100 people in the standby queue, numeric}
}
}
\source{
"entities.csv" Disney World Ride Wait Time Datasets, TouringPlans.com, September 2021,
\url{https://touringplans.com/walt-disney-world/crowd-calendar#DataSets}, Accessed 16 September 2021
}
\usage{
attractions
}
\description{
A dataset containing data about Disney theme park attractions.
}
\keyword{datasets}
