% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinosaur.R
\docType{data}
\name{dinosaur}
\alias{dinosaur}
\title{DINOSAUR Wait Times}
\format{
A data frame with 213,381 rows and 4 variables:
\describe{
\item{date}{Park Day (not actual date stamp of the wait time, since some are after midnight), Date}
\item{datetime}{date-time stamp of wait time,	YYYY-MM-DD HH:MM:SS}
\item{sactmin}{Actual Wait Time (in minutes), numeric}
\item{spostmin}{Standby Posted Wait Time (in minutes),	numeric}
}
}
\source{
"dinosaur.csv" Disney World Ride Wait Time Datasets, TouringPlans.com, September 2021,
\url{https://touringplans.com/walt-disney-world/crowd-calendar#DataSets}, Accessed 16 September 2021
}
\usage{
dinosaur
}
\description{
A dataset containing DINOSAUR wait times.
}
\keyword{datasets}
