% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kilimanjaro_safaris.R
\docType{data}
\name{kilimanjaro_safaris}
\alias{kilimanjaro_safaris}
\title{Kilimanjaro Safaris Wait Times}
\format{
A data frame with 257,785 rows and 4 variables:
\describe{
\item{park_date}{Park Day (not actual date stamp of the wait time, since some are after midnight), Date}
\item{wait_datetime}{date-time stamp of wait time,	YYYY-MM-DD HH:MM:SS}
\item{wait_minutes_actual}{Actual Wait Time (in minutes), numeric}
\item{wait_minutes_posted}{Standby Posted Wait Time (in minutes),	numeric}
}
}
\source{
"kilimanjaro_safaris.csv" Disney World Ride Wait Time Datasets, TouringPlans.com, August 2023,
\url{https://touringplans.com/walt-disney-world/crowd-calendar#DataSets}, Accessed 28 August 2023
}
\usage{
kilimanjaro_safaris
}
\description{
A dataset containing Kilimanjaro Safaris wait times.
}
\keyword{datasets}
