% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirates_of_caribbean.R
\docType{data}
\name{pirates_of_caribbean}
\alias{pirates_of_caribbean}
\title{Pirates of the Caribbean Wait Times}
\format{
A data frame with 255,429 rows and 4 variables:
\describe{
\item{date}{Park Day (not actual date stamp of the wait time, since some are after midnight), Date}
\item{datetime}{date-time stamp of wait time,	YYYY-MM-DD HH:MM:SS}
\item{wait_minutes_actual}{Actual Wait Time (in minutes), numeric}
\item{wait_minutes_posted}{Standby Posted Wait Time (in minutes),	numeric}
}
}
\source{
"pirates_of_caribbean.csv" Disney World Ride Wait Time Datasets, TouringPlans.com, August 2023,
\url{https://touringplans.com/walt-disney-world/crowd-calendar#DataSets}, Accessed 28 August 2023
}
\usage{
pirates_of_caribbean
}
\description{
A dataset containing Pirates of the Caribbean wait times.
}
\keyword{datasets}
