% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourinplans_datasets.R
\docType{data}
\name{touringplans_datasets}
\alias{touringplans_datasets}
\title{Metadata about attractions included in datasets}
\format{
A data frame with 14 rows and 7 variables:
\describe{
\item{dataset_name}{Name of dataset for ride wait times in this package, alphanumeric}
\item{name}{Disney's official attraction name, alphanumeric}
\item{short_name}{Short name, alphanumeric}
\item{park}{Disney theme park where the attraction is located}
\item{land}{Section of the park an attraction is located}
\item{opened_on}{Attraction's opening date, date}
\item{duration}{Duration of attraction in minutes, numeric}
}
}
\source{
"entities.csv" Disney World Ride Wait Time Datasets, TouringPlans.com, September 2021,
\url{https://touringplans.com/walt-disney-world/crowd-calendar#DataSets}, Accessed 16 September 2021
}
\usage{
touringplans_datasets
}
\description{
A dataset containing data about Disney theme park attractions.
}
\keyword{datasets}
