% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touringplans_metadata.R
\docType{data}
\name{touringplans_metadata}
\alias{touringplans_metadata}
\title{Meta data for Disney Touring Plan Data}
\format{
A data frame with 2079 rows and 181 variables:
\describe{
\item{DATE}{Park Day (not actual date stamp of the wait time, since some are after midnight), Date}
\item{WDW_TICKET_SEASON}{Walt Disney World Single Day Price Type, alphanumeric}
\item{DAYOFWEEK}{Day of Week, numeric}
\item{DAYOFYEAR}{Day of Year (0 to 365), numeric}
\item{WEEKOFYEAR}{Week of Year (0 to 53), numeric}
\item{MONTHOFYEAR}{Month of Year (1 to 12), numeric}
\item{YEAR}{Year, numeric}
\item{SEASON}{Seasonal Category (Fall, Winter, Thanksgiving, etc.), alphanumeric}
\item{HOLIDAYPX}{Proximity to Holiday (2-directional) (in days), numeric}
\item{HOLIDAYM}{Holiday Metric (1 to 5 ranking), numeric}
\item{HOLIDAYN}{Name of Holiday, alphanumeric}
\item{HOLIDAY}{Day is a Holiday, Boolean}
\item{WDWTICKETSEASON}{WDW 1-day ticket season, alphanumeric}
\item{WDWRaceN}{XXX Event Code, alphanumeric}
\item{WDWeventN}{XXX Event Code, alphanumeric}
\item{WDWevent}{WDW Event, Boolean}
\item{WDWrace}{WDW Race, Boolean}
\item{WDWSEASON}{WDW Season, alphanumeric}
\item{WDWMAXTEMP}{WDW Max Temperature, numeric}
\item{WDWMINTEMP}{WDW Min Temperature, numeric}
\item{WDWMEANTEMP}{WDW Average Temperature, numeric}
\item{MKeventN}{Magic Kingdom Event Code, alphanumeric}
\item{MKevent}{Magic Kingdom Event, Boolean}
\item{EPeventN}{Epcot Event Code, alphanumeric}
\item{EPevent}{Epcot Event, Boolean}
\item{HSeventN}{Hollywood Studios Event Code, alphanumeric}
\item{HSevent}{Hollywood Studios Event, Boolean}
\item{AKeventN}{Animal Kingdom Event Code, alphanumeric}
\item{AKevent}{Animal Kingdom Event, Boolean}
\item{HOLIDAYJ}{Name of Jewish Holiday, alphanumeric}
\item{INSESSION}{Percentage of Schools in Session, numeric}
\item{INSESSION_ENROLLMENT}{Percentage of Schools in Session Weighted by Enrollment, numeric}
\item{INSESSION_WDW}{Percentage of Schools in Session WDW markets only, numeric}
\item{INSESSION_DLR}{Percentage of Schools in Session DLR markets only, numeric}
\item{INSESSION_SQRT_WDW}{Percentage of Schools in Session Weighted by Square Root of Distance to WDW, numeric}
\item{INSESSION_SQRT_DLR}{Percentage of Schools in Session Weighted by Square Root of Distance to DLR, numeric}
\item{INSESSION_CALIFORNIA}{Percentage of Schools in Session California Only, numeric}
\item{INSESSION_DC}{Percentage of Schools in Session D.C. Only, numeric}
\item{inSession_Central_FL}{Percentage of Schools in Session Within Central Florida Only, numeric}
\item{INSESSION_DRIVE1_FL}{Percentage of Schools in Session Within Driving Distance to Florida Only, numeric}
\item{INSESSION_DRIVE2_FL}{Percentage of Schools in Session Within Driving Distance to Florida Only, numeric}
\item{INSESSION_DRIVE_CA}{Percentage of Schools in Session Within Driving Distance to California Only, numeric}
\item{INSESSION_FLORIDA}{Percentage of Schools in Session Florida Only, numeric}
\item{INSESSION_MARDI_GRAS}{Percentage of Schools in Session Mardi Gras States Only, numeric}
\item{INSESSION_MIDWEST}{Percentage of Schools in Session Midwest Only, numeric}
\item{INSESSION_NY_NJ}{Percentage of Schools in Session New York and New Jersey Only, numeric}
\item{INSESSION_NY_NJ_PA}{Percentage of Schools in Session New York, New Jersey and Pennsylvania Only, numeric}
\item{INSESSION_NEW_ENGLAND}{Percentage of Schools in Session New England Only, numeric}
\item{INSESSION_NEW_JERSEY}{Percentage of Schools in Session New Jersey Only, numeric}
\item{INSESSION_NOTHWEST}{Percentage of Schools in Session Northwest Only, numeric}
\item{INSESSION_PLANES}{Percentage of Schools in Session Planes Only, numeric}
\item{INSESSION_SOCAL}{Percentage of Schools in Session Southern California Only, numeric}
\item{INSESSION_SOUTHWEST}{Percentage of Schools in Session Southwest Only, numeric}
\item{SUNSET_WDW}{Time of Sunset in Orlando, HH:MM}
\item{MKEMHMORN}{Magic Kingdom Extra Magic Hour Morning, Boolean}
\item{MKEMHMYEST}{Magic Kingdom Extra Magic Hour Morning Yesterday, Boolean}
\item{MKEMHMTOM}{Magic Kingdom Extra Magic Hour Morning Tomorrow, Boolean}
\item{MKEMHEVE}{Magic Kingdom Extra Magic Hour Evening, Boolean}
\item{MKHOURSEMH}{Total Opening Hours including EMH for Magic Kingdom, numeric}
\item{MKHOURSEMHYEST}{Yesterday\'s Total Opening Hours including EMH for Magic Kingdom, numeric}
\item{MKHOURSEMHTOM}{Tomorrow\'s Total Opening Hours including EMH for Magic Kingdom, numeric}
\item{MKEMHEYEST}{Magic Kingdom Extra Magic Hour Evening Yesterday, Boolean}
\item{MKEMHETOM}{Magic Kingdom Extra Magic Hour Evening Tomorrow, Boolean}
\item{EPEMHMORN}{Epcot Extra Magic Hour Morning, Boolean}
\item{EPEMHMYEST}{Epcot Extra Magic Hour Morning Yesterday, Boolean}
\item{EPEMHMTOM}{Epcot Extra Magic Hour Morning Tomorrow, Boolean}
\item{EPEMHEVE}{Epcot Extra Magic Hour Evening, Boolean}
\item{EPEMHEYEST}{Epcot Extra Magic Hour Evening Yesterday, Boolean}
\item{EPEMHETOM}{Epcot Extra Magic Hour Evening Tomorrow, Boolean}
\item{EPHOURSEMH}{Total Opening Hours including EMH for Epcot, numeric}
\item{EPHOURSEMHYEST}{Yesterday\'s Total Opening Hours including EMH for Epcot, numeric}
\item{EPHOURSEMHTOM}{Tomorrow\'s Total Opening Hours including EMH for Epcot, numeric}
\item{HSEMHMORN}{Hollywood Studios Extra Magic Hour Morning, Boolean}
\item{HSEMHMYEST}{Hollywood Studios Extra Magic Hour Morning Yesterday, Boolean}
\item{HSEMHMTOM}{Hollywood Studios Extra Magic Hour Morning Tomorrow, Boolean}
\item{HSEMHEVE}{Hollywood Studios Extra Magic Hour Evening, Boolean}
\item{HSEMHEYEST}{Hollywood Studios Extra Magic Hour Evening Yesterday, Boolean}
\item{HSEMHETOM}{Hollywood Studios Extra Magic Hour Evening Tomorrow, Boolean}
\item{HSHOURSEMH}{Total Opening Hours including EMH for Hollywood Studios, numeric}
\item{HSHOURSEMHYEST}{Yesterday\'s Total Opening Hours including EMH for Hollywood Studios, numeric}
\item{HSHOURSEMHTOM}{Tomorrow\'s Total Opening Hours including EMH for Hollywood Studios, numeric}
\item{AKEMHMORN}{Animal Kingdom Extra Animal Hour Morning, Boolean}
\item{AKEMHMYEST}{Animal Kingdom Extra Animal Hour Morning Yesterday, Boolean}
\item{AKEMHMTOM}{Animal Kingdom Extra Animal Hour Morning Tomorrow, Boolean}
\item{AKEMHEVE}{Animal Kingdom Extra Animal Hour Evening, Boolean}
\item{AKEMHEYEST}{Animal Kingdom Extra Animal Hour Evening Yesterday, Boolean}
\item{AKEMHETOM}{Animal Kingdom Extra Animal Hour Evening Tomorrow, Boolean}
\item{AKHOURSEMH}{Total Opening Hours including EMH for Animal Kingdom, numeric}
\item{AKHOURSEMHYEST}{Yesterday\'s Total Opening Hours including EMH for Animal Kingdom, numeric}
\item{AKHOURSEMHTOM}{Tomorrow\'s Total Opening Hours including EMH for Animal Kingdom, numeric}
\item{MKOPEN}{Opening Hour for Magic Kingdom, HH:MM}
\item{MKCLOSE}{Closing Hour for Magic Kingdom, HH:MM}
\item{MKHOURS}{Total Opening Hours for Magic Kingdom, numeric}
\item{MKEMHOPEN}{Magic Kingdom Extra Magic Hour Opening Time, HH:MM}
\item{MKEMHCLOSE}{Magic Kingdom Extra Magic Hour Closing Time, HH:MM}
\item{MKOPENYEST}{Yesterday\'s Opening Hour for Magic Kingdom, HH:MM}
\item{MKCLOSEYEST}{Yesterday\'s Closing Hour for Magic Kingdom, HH:MM}
\item{MKHOURSYEST}{Yesterday\'s Total Opening Hours for Magic Kingdom, numeric}
\item{MKOPENTOM}{Tomorrow\'s Opening Hour for Magic Kingdom, HH:MM}
\item{MKCLOSETOM}{Tomorrow\'s Closing Hour for Magic Kingdom, HH:MM}
\item{MKHOURSTOM}{Tomorrow\'s Total Opening Hours for Magic Kingdom, numeric}
\item{EPOPEN}{Opening Hour for Epcot, HH:MM}
\item{EPCLOSE}{Closing Hour for Epcot, HH:MM}
\item{EPHOURS}{Total Opening Hours for Epcot, numeric}
\item{EPEMHOPEN}{Epcot Kingdom Extra Magic Hour Opening Time, HH:MM}
\item{EPEMHCLOSE}{Epcot Extra Magic Hour Closing Time, HH:MM}
\item{EPOPENYEST}{Yesterday\'s Opening Hour for Epcot, HH:MM}
\item{EPCLOSEYEST}{Yesterday\'s Closing Hour for Epcot, HH:MM}
\item{EPHOURSYEST}{Yesterday\'s Total Opening Hours for Epcot, numeric}
\item{EPOPENTOM}{Tomorrow\'s Opening Hour for Epcot, HH:MM}
\item{EPCLOSETOM}{Tomorrow\'s Closing Hour for Epcot, HH:MM}
\item{EPHOURSTOM}{Tomorrow\'s Total Opening Hours for Epcot, numeric}
\item{HSOPEN}{Opening Hour for Disney Hollywood Studios, HH:MM}
\item{HSCLOSE}{Closing Hour for Disney Hollywood Studios, HH:MM}
\item{HSHOURS}{Total Opening Hours for Disney Hollywood Studios, numeric}
\item{HSEMHOPEN}{Hollywood Studios Extra Magic Hour Opening Time, HH:MM}
\item{HSEMHCLOSE}{Hollywood Studios Extra Magic Hour Closing Time, HH:MM}
\item{HSOPENYEST}{Yesterday\'s Opening Hour for Hollywood Studios, HH:MM}
\item{HSCLOSEYEST}{Yesterday\'s Closing Hour for Hollywood Studios, HH:MM}
\item{HSHOURSYEST}{Yesterday\'s Total Opening Hours for Hollywood Studios, numeric}
\item{HSOPENTOM}{Tomorrow\'s Opening Hour for Hollywood Studios, HH:MM}
\item{HSCLOSETOM}{Tomorrow\'s Closing Hour for Hollywood Studios, HH:MM}
\item{HSHOURSTOM}{Tomorrow\'s Total Opening Hours for Hollywood Studios, numeric}
\item{AKOPEN}{Opening Hour for Animal Kingdom, HH:MM}
\item{AKCLOSE}{Closing Hour for Animal Kingdom, HH:MM}
\item{AKHOURS}{Total Opening Hours for Animal Kingdom, numeric}
\item{AKEMHOPEN}{Animal Kingdom Extra Magic Hour Opening Time, HH:MM}
\item{AKEMHCLOSE}{Animal Kingdom Extra Magic Hour Closing Time, HH:MM}
\item{AKOPENYEST}{Yesterday\'s Opening Hour for Animal Kingdom, HH:MM}
\item{AKCLOSEYEST}{Yesterday\'s Closing Hour for Animal Kingdom, HH:MM}
\item{AKHOURSYEST}{Yesterday\'s Total Opening Hours for Animal Kingdom, numeric}
\item{AKOPENTOM}{Tomorrow\'s Opening Hour for Animal Kingdom, HH:MM}
\item{AKCLOSETOM}{Tomorrow\'s Closing Hour for Animal Kingdom, HH:MM}
\item{AKHOURSTOM}{Tomorrow\'s Total Opening Hours for Animal Kingdom, numeric}
\item{MKDAYSBEFORENONEVENT}{Number of Days before next non-event day, numeric}
\item{MKDAYSSINCENONEVENT}{Number of Days since the last non-event day, numeric}
\item{MKEVENTSTREAK}{Number of Days of consecutive party days, numeric}
\item{MKEVENTSTREAK_F}{Number of days of future consecutive party days, numeric}
\item{PARTYSEASON_WDW}{Season of Magic Kingdom Parties, alphanumeric}
\item{WDWMINTEMP_MEAN}{Average minimum daily temperature, numeric}
\item{WEATHER_WDWHIGH}{Historical High Temperature, numeric}
\item{WEATHER_WDWLOW}{Historical Low Temperature, numeric}
\item{WEATHER_WDWPRECIP}{Historical Precipitation, numeric}
\item{CAPACITYLOST_MK}{Total hourly capacity lost on that park day (due to attraction closures), numeric}
\item{CAPACITYLOST_EP}{Total hourly capacity lost on that park day, numeric}
\item{CAPACITYLOST_HS}{Total hourly capacity lost on that park day, numeric}
\item{CAPACITYLOST_AK}{Total hourly capacity lost on that park day, numeric}
\item{CAPACITYLOSTWGT_MK}{Total hourly capacity lost on that park day, weighted by attraction popularity, numeric}
\item{CAPACITYLOSTWGT_EP}{Total hourly capacity lost on that park day, weighted by attraction popularity, numeric}
\item{CAPACITYLOSTWGT_HS}{Total hourly capacity lost on that park day, weighted by attraction popularity, numeric}
\item{CAPACITYLOSTWGT_AK}{Total hourly capacity lost on that park day, weighted by attraction popularity, numeric}
\item{EP09CAPACITY}{Hourly capacity of Soarin\', numeric}
\item{HS20CAPACITY}{Hourly capacity of Toy Story Mania, numeric}
\item{MKPRDDAY}{Number of Daytime Parades at Magic Kingdom, numeric}
\item{MKPRDDT1}{1st Parade Time at Magic Kingdom, HH:MM}
\item{MKPRDDT2}{2nd Parade Time at Magic Kingdom, HH:MM}
\item{MKPRDDN}{Magic Kingdom Day Time Parade Name, alphanumeric}
\item{MKPRDNGT}{Number of Night Time Parades at Magic Kingdom, Boolean}
\item{MKPRDNT1}{1st Night Parade Time at Magic Kingdom, HH:MM}
\item{MKPRDNT2}{2nd Night Parade Time at Magic Kingdom, HH:MM}
\item{MKPRDNN}{Magic Kingdom Night Parade Name, alphanumeric}
\item{MKFIREWK}{Number of Fireworks Shows at Magic Kingdom, numeric}
\item{MKFIRET1}{1st Magic Kingdom Fireworks time, HH:MM}
\item{MKFIRET2}{2nd Magic Kingdom Fireworks time, HH:MM}
\item{MKFIREN}{Evening Fireworks Name at Magic Kingdom, alphanumeric}
\item{EPFIREWK}{Number of Fireworks Shows at Epcot, numeric}
\item{EPFIRET1}{1st Epcot Fireworks time, HH:MM}
\item{EPFIRET2}{2nd Epcot Fireworks time, HH:MM}
\item{EPFIREN}{Evening Fireworks Name at Epcot, alphanumeric}
\item{HSPRDDAY}{Number of Daytime Parades at Hollywood Studios, numeric}
\item{HSPRDDT1}{1st Parade Time at Hollywood Studios, HH:MM}
\item{HSPRDDN}{Hollywood Studios Day Time Parade Name, alphanumeric}
\item{HSFIREWK}{Number of Fireworks Shows at Hollywood Studios, numeric}
\item{HSFIRET1}{1st Hollywood Studios Fireworks time, HH:MM}
\item{HSFIRET2}{2nd Hollywood Studios Fireworks time, HH:MM}
\item{HSFIREN}{Evening Fireworks Name at Hollywood Studios, alphanumeric}
\item{HSSHWNGT}{Number of Night Time Shows at Hollywood Studios, numeric}
\item{HSSHWNT1}{1st Hollywood Studios Night Time Show time, HH:MM}
\item{HSSHWNT2}{2nd Hollywood Studios Night Time Show time, HH:MM}
\item{HSSHWNN}{Night Time Show Name at Hollywood Studios, alphanumeric}
\item{HSFIREWKS}{Number of Fireworks Shows at Hollywood Studios, numeric}
\item{AKPRDDAY}{Number of Daytime Parades at Animal Kingdom, numeric}
\item{AKPRDDT1}{1st Parade Time at Animal Kingdom, HH:MM}
\item{AKPRDDT2}{2nd Parade Time at Animal Kingdom, HH:MM}
\item{AKPRDDN}{Animal Kingdom Day Time Parade Name, alphanumeric}
\item{AKFIREN}{Fireworks Name at Animal Kingdom, alphanumeric}
\item{AKSHWNGT}{Number of Night Time Shows at Animal Kingdom, numeric}
\item{AKSHWNT1}{1st Animal Kingdom Night Time Show time, HH:MM}
\item{AKSHWNT2}{2nd Animal Kingdom Night Time Show time, HH:MM}
\item{AKSHWNN}{Night Time Show Name at Animal Kingdom, alphanumeric}
}
}
\source{
"metadata.csv" Disney World Ride Wait Time Datasets, TouringPlans.com, September 2021,
\url{https://touringplans.com/walt-disney-world/crowd-calendar#DataSets}, Accessed 16 September 2021
}
\usage{
touringplans_metadata
}
\description{
A dataset containing daily details on Disney World Parks
}
\keyword{datasets}
