% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_story_mania.R
\docType{data}
\name{toy_story_mania}
\alias{toy_story_mania}
\title{Toy Story Mania! Wait Times}
\format{
A data frame with 240,079 rows and 4 variables:
\describe{
\item{date}{Park Day (not actual date stamp of the wait time, since some are after midnight), Date}
\item{datetime}{date-time stamp of wait time,	YYYY-MM-DD HH:MM:SS}
\item{SACTMIN}{Actual Wait Time (in minutes), numeric}
\item{SPOSTMIN}{Standby Posted Wait Time (in minutes),	numeric}
}
}
\source{
"toy_story_mania.csv" Disney World Ride Wait Time Datasets, TouringPlans.com, September 2021,
\url{https://touringplans.com/walt-disney-world/crowd-calendar#DataSets}, Accessed 16 September 2021
}
\usage{
toy_story_mania
}
\description{
A dataset containing Toy Story Mania! wait times.
}
\keyword{datasets}
